requirements.txt
customtkinter==5.0.3
matplotlib==3.5.2
numpy==1.22.4
Pillow==9.5.0
pygame==2.1.2
typing_extensions==4.5.0
pyperclip==1.8.2
